[TASK] Add close confirmation modal to scheduler
Currently, if a task was changed in the edit/add view
the close button goes back to list view without
letting the user know about potentially lost changes.

This has been changed, by adding a confirmation modal
when the close button was clicked but changes are not
yet saved.

Resolves: #100524
Releases: main
Change-Id: I623a53047605bc85046cbd35a3e9a61c2e21b508
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78532
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>