[BUGFIX] Use IntersectionObserver to determine if a tree is visible
The problem with the current implementation is that it does not
consider that a tab may already be active but not visible initially.
In that case the `shown.bs.tab` event is never triggered.

We are moving to an observer that stays active until the element
is visible once, after that the observer for the tree is de-registered.

Resolves: #97082
Releases: main, 11.5
Change-Id: If63534ac28afd7d7ca53a963f6fad34ae9522885
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78556
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>