[BUGFIX] ckeditor: Do not update source element on document change
The change in #99669 introduced an update of the source element bound to
the ckeditor instance. Unfortunately, this leads to a recursion loop
when the "Source view" is active and the save action in FormEngine is
invoked.

Apparently, updating the source element is not required and is therefore
removed again.

Resolves: #100557
Related: #99669
Releases: main
Change-Id: I84434bade2a2512f41a19dd56eda99ac5e8f41b0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78593
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>