[BUGFIX] Ensure email links do not throw an exception
When upgrading a project to TYPO3 v12,
the new email link validator now
checks for other properties than the
email address such as "to" etc.

However, in previous TYPO3 v11 installations,
it was possible to enter a link like " benni@typo3.org"
and it worked, the new link handler now fatals
due a stricter parsing, which is now relaxed,
as the value is now trimmed.

Once editors modify all their links, it would
work, however, that's not in scope for the
LTS, and would require a Upgrade Wizard
which also checks for HTML links. This way,
it is much easier to overcome the problem
by trimming the values.

Resolves: #100568
Releases: main
Change-Id: I65156826aa7944a7d4af9ae3c830012d228b12c2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78603
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>