[TASK] Reorganize & streamline sudo mode handling
Issue #92836 introduced the "sudo mode" to all admin tool modules,
but with using the session storage of the install tool (based on
PHP native `$_SESSION` handling) - which is different to the usual
behavior and session handling in the TYPO3 backend.

This change changes the session handling to backend user sessions,
add a generic way to configure "sudo mode" per backend routes (not
hardcoding its invocation in the admin tool controller anymore).

In addition, "sudo mode" is now enabled for the extension manager
module as well - this was not possible before, due to the hardwired
implementation to `ext:install`.

In developer context, "sudo mode" is granted automatically (#93160).

Resolves: #94246
Releases: main
Change-Id: I4fee2b7da3da25c8c6479eab7478b38cdc3b5a4d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78546
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>