[TASK] Streamline media queries and breakpoints
The Bootstrap CSS Framework changed the handling of media queries
with version 4.6 but the core never adapted to this. We are now
finishing the migration and dropping all `$screen-` variables.

All media queries that are bound to bootstrap breakpoints are
now using the provided mixins.

In addition the unused `.dl-horizontal` was dropped.

Resolves: #100588
Releases: main
Change-Id: I2d47d62f4d110a290607708ebb496dc23249ee17
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78630
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>