[TASK] Drop a.disabled styling
The special handling for a.disabled is obsolete, is was introduced
to prevent the default link handling for the new content element
wizard links while the component was not loaded.

The wizard button is now a webcomponent and then handling
and special styling is already handled through the boostrap
css framework.

The styling can be dropped.

Resolves: #100592
Releases: main
Change-Id: Icca02cc9f1cc84c19681149e764854fe426d103f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78635
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>