[TASK] Do not set &nbsp; as label for checkboxes in debug mode
There is no need to add a non-breaking space as a fallback if the label
is empty for checkboxes. This results in unnecessary occupied
space. We prefer just an empty string.

To ensure we only occupy the space we need to, we are trimming
unnecessary whitespaces from the final label.

Resolves: #100594
Releases: main
Change-Id: I941a197dc9e756c978b19dc8275449fcb1874c31
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78637
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>