[TASK] Deprecate GeneralUtility::_GET()
This deprecates GeneralUtility::_GET() as last
member of the get/post related family, next to
GU::_POST(), GU::_GP() and GU::_GPmerged(): The
request object has to be used instead.

This patch is the end of a journey that took
us various years. The last main one left is
GeneralUtility::getIndpEnv(), but that's a
story for TYPO3 v13.

Resolves: #100596
Related: #99633
Related: #99615
Related: #100053
Related: #100584
Releases: main
Change-Id: I591dd289289f5e4d7d35b74f853b067fbe975624
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78639
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>