[TASK] Update display of 'Create multiple pages' wizard
The markup of the form to create multiple pages was updated to match
the regular markup.

After creation, the new and existing pages are now displayed in
a table. The displayed page entries now contain a context menu and
an edit button to provide a possibility to jump directly to the
page settings.

Resolves: #100604
Releases: main
Change-Id: Ib4377c99f4d0a78de37df9c2f7cc56384f696f82
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78645
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>