[BUGFIX] Extbase BackendConfigurationManager fakes sys_template
The TypoScript parser has to load TypoScript from globals
($GLOBALS['TYPO3_CONF_VARS']['FE']['defaultTypoScript_')
and TypoScript from extension static files
(ext_typoscript_*.typoscript) at some point.

It usually adds this before the first "root" sys_template
row. This is fine in frontend.

Extbase backend modules rely on frontend TypoScript for
configuration as well. This decision from the dark ages
hurts us plenty, but is very hard to get rid of.

The extbase BackendConfigurationManager does a lot of
guesswork to calculate such frontend TypoScript: There
are various special cases, especially when an extbase
backend module is not loaded within page context.

The ext:form extension triggers such a special case:
When there is only a single page marked as "is_siteroot"
without a sys_template record, configuration from
globals is not included. The configuration is then
incomplete and the module fails.

The old TypoScript parser had a dedicated hack and
toggle for this, which forced it to still include
global configuration. The new TypoScript parser
does not contain this hack. Instead, the extbase
BackendConfigurationManager "fakes" a sys_template
row to trigger the global inclusion. This works if
there is no "is_siteroot" page at all, but is not
triggered when a page could be determined.

The patch slightly changes the extbase logic to also
add the fake row if a page could be determined, but
no sys_template record exists.

Additionally, the ext:form backend controllers now
throw a dedicated exception if configuration is
incomplete.

Resolves: #99458
Related: #97816
Releases: main
Change-Id: Ie48cd5c97704c9c4a0ad7a5e555f49f181f88006
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78650
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Kevin Appelt <kevin.appelt@icloud.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Kevin Appelt <kevin.appelt@icloud.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>