[BUGFIX] Adjust secondary color for components that respect prefers-color-scheme
With the migration from `text-muted` to `text-body-secondary`
the descriptive text for clearing cache actions became unreadable
in darkmode.

The dropdown component is already respecting the preferred
color scheme of the user. Since we have no global support for
darkmode and only have some components react, we need to
adjust these manually for now.

Resolves: #100628
Releases: main
Change-Id: Iffffdd7675dc28c026cb650a0897b921c2fbc30f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78676
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>