[BUGFIX] `ceil()` dimensions in CropperEvent
The dimensions available in `CropperEvent` are fractions in some cases,
that probably happens when an image gets upscaled. This causes issues
as those dimensions are apparently adjusted downward in the cropper
library, resulting in dimension reduced by 1px. To solve this issue,
`Math.ceil()` is applied to round up incoming dimensions.

Resolves: #96722
Releases: main, 11.5
Change-Id: I537bccd32a634805a37957bea59ec51e5e856272
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78717
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>