[BUGFIX] Pass markup as fragment in `debugInPopUpWindow()`
The method `DebugUtility::debugInPopUpWindow()` renders a modal with
a debug dump, if applicable. The content will always contain markup,
which is rendered as plain text in the modal, was intended by the modal
component.

To have the dump correctly rendered, the markup is passed as a fragment
to the modal component.

Resolves: #100652
Releases: main
Change-Id: I1cd1a97a5cd794e941f1a5830d8fd0cd97d8b1cb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78718
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>