[BUGFIX] Properly check file is missing to prevent exception
In case a file is still indexed but missing on
the driver, searching for the file in the filelist
module previously led to an exception.

This is now fixed by checking whether the file
is missing before fetching its parent folder,
which is used to generate the full file path.

Resolves: #100689
Releases: main
Change-Id: I3ea5ec09be366c911dbf5e8217d4818e708c5c82
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78773
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jochen <rothjochen@gmail.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>