[BUGFIX] Handle dynamic CSP nonce values during frontend rendering
The CSP nonce value was not assigned in frontend rendering process.
Besides that, the value would have been cached statically, which is
not in line with the concept of a "number used once (nonce)".

A new `INT_SCRIPT´ instruction is added, which permanently post
processes cached data and replaces the cached nonce with the
dynamic value of the corresponding actual request.

Resolves: #100665
Releases: main
Change-Id: I52107167541993112c9abe331bd00c86726fc5fa
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78660
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>