[BUGFIX] Correctly invoke delete and mute actions of CSP module
The `deleteAction` and `muteAction` of the CSP module endpoint
expected to receive multiple `summaries` as identifiers, however
just a single `summary` was sent, and the corresponding action
was not executed on the server side.

Besides that, a new scope specific "remove all" action has been
added to the CSP module and endpoint.

Resolves: #100618
Releases: main
Change-Id: I9731cfd2e1afcccd70d2d3df2e7da2fa1f4a87e3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78793
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>