[BUGFIX] Use `importShim()` to load init modules
With #100688, the module `es-module-shims` was updated. Unfortunately,
with 1.5.3 an incompatibility with Firefox < 108 was introduced,
affecting current Firefox ESR releases (currently at 102, at the time
of writing this patch).

To workaround this issue, the init scripts used in the Install Tool now
use `importShim()` for the time being, until Firefox ESR hits version
108.

Resolves: #100756
Related: #100688
Releases: main
Change-Id: I1337ebbf1f01db287a53ff7ccad50cb4631250d9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78881
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>