[BUGFIX] Pass the full page record to the ConditionMatcher
The PageLayoutResolver, which is used for
"tree.pagelayout" TypoScript conditions
requires the passed page record to contain
the "backend_layout" field value.

Since this field is by default not a
rootline field, the TSFE->page array
is now passed again instead of the
shrinked record from the rootline.

Resolves: #100764
Related: #100047
Releases: main
Change-Id: I28ea56fb70b026e6d88b1cb00838490b35c2d5d4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78891
Tested-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Jochen <rothjochen@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>