[BUGFIX] Allow ViewHelpers from libraries
With TYPO3 11 it was possible to use ViewHelpers from composer libraries
that aren't TYPO3 extensions and thus don't use Symfony dependency
injection. With the removal of objectManager, this doesn't work anymore
in TYPO3 12 because the dependency injection can only instantiate
ViewHelper classes that are defined as public services explicitly.

This change reintroduces that functionality by falling back to "new" if
the class exists but isn't registered as a service. Fluid standalone
uses the same method to instantiate a ViewHelper.

Resolves: #100785
Releases: main, 12.4
Change-Id: I495e006fa15eb815fb3734b4c73361821869f4ad
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78917
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>