[BUGFIX] Change selector to get container in `files` element
The identifier used to get a container in `type=files` based on
`querySelector()` which flawed if the identifier contained a dot.
The example identifier "#foo-data-foo.bar" resulted in "find an
element with id 'foo-data-foo', having the class 'bar'", which
is wrong.

To solve this issue, while still limiting the scope to the web
component, the selector is changed to use the `[id='']` syntax,
which doesn't show the erroneous behavior.

Resolves: #100777
Releases: main, 12.4
Change-Id: I9edff128c612cc6a614b65866533c3fb1c731f4e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78943
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>