[BUGFIX] Update `symfony/property-info` to v6.2.11
Handling of collections without a dedicated type has changed in
`symfony/property-info` with v6.2.11, [1] seems to be the offensive
commit.

This should be backward compatible, as only string|int make
sense for array keys and therefore is the same for array-like
object annotation syntax.

We raise the version constraint for `symfony/property-info`
package to a slightly higher patchlevel version to ensure
consistent behaviour. This is done for shipped version and
the core development, therefore the raises are also added
to the corresponding extension composer.json files.

Used command(s):

> composer require --no-update \
    -d typo3/sysext/extbase \
    "symfony/property-info":"^6.2.11"

> composer require \
    "symfony/property-info":"^6.2.11"

[1] https://github.com/symfony/symfony/pull/49557

Resolves: #100927
Releases: main, 12.4
Change-Id: I1465206414cd7e596d0852ed4288c5013db23c2d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79151
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>