[TASK] Remove obsolete `composer/package-versions-deprecated`
Package `composer/package-versions-deprecated` is required
as development dependency in the root composer.json. That
package itself is deprecated and it is recommended to use
class `Composer\InstalledVersions` directly with composer 2+
only support.

TYPO3 requires composer v2.x as a minimum since TYPO3 v11,
so there is no need to keep this development dependency
installed. It's not used anyway in the development cycle.

Used command(s):

> composer remove --dev "composer/package-versions-deprecated"
> composer config --unset \
    allow-plugins.composer/package-versions-deprecated

Resolves: #100942
Releases: main, 12.4
Change-Id: I9d0bb62e8e3f8e7aa5f33f3e3f49a79695ba3244
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79161
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>