[BUGFIX] Flush fluid cache when switching debug output in AdminPanel
Currently, when switching the 'Show fluid debug output' in AdminPanel,
the relevant frontend caches weren't flushed, causing pages to be
rendered using the old ViewHelper namespace configuration, so that
the debug viewhelpers weren't added/removed, depending on the value.

This commit makes the caches get flushed when the option is changed.

Resolves: #99883
Releases: main, 12.4
Change-Id: I495d678bce2128f538b46f79485fdab67581a29f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79402
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Frank Nägler <frank.naegler@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Frank Nägler <frank.naegler@typo3.com>