[BUGFIX] Enable client side min/max validation for type folder
While separating type folder from type group, it was forgotten to enable
client-side validation for min/max checks in the FormEngine validation.

Furthermore, it is recommended to use minitems over required=true for
fields with multiple items. Reasons are: 1. it is not documented, 2.
to better conform with the counterpart option maxitems.

Resolves: #101089
Related: #96983
Releases: main, 12.4
Change-Id: I4a8d01e15bb72ceff8130c1dcc18b829097375ac
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79421
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: core-ci <typo3@b13.com>