[TASK] Add shared cache for VariableProcessor
VariableProcessor::addHash() is called to convert URI path
variables that don't comply with certain rules to a hash,
which does.

First of all, variables must not exceed a certain length
of 32, second of all, variables must only contain word
characters aka "\w" or "[A-Za-z0-9_]".

So whenever a variable is too long or contains invalid
characters, it is converted to an md5 hash which complies
with both rules.

It has been reported that the checks, whether to generate a
hash and the hashing would have a very noticable performance
impact.

This change introduces as specific runtime cache for
VariableProcessor that is shared with any other new instance.
This way, the necessity whether a hash is required and the hash
representation of a value can be resolved from this cache.

For instance, this would speed up scenarios that are enerating
large product lists and generate, when every invocation with a
new UID would have triggered a new hash check and assignment.

Resolves: #100974
Releases: main, 12.4
Change-Id: I19e1d138a79792cc05a6a908f1ee5f87236a65bd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79522
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>