[BUGFIX] Respect preferred browser language in backend login
With #100255 it has been ensured, that the translate ViewHelper uses
the site language locale. For backend context, the locale is
resolved from `$GLOBALS['BE_USER']->user['lang']`.

In case of the backend login or password reset form , the evaluated
preferred browser language is however not set to the global `BE_USER`
object, when no user is logged in. As a result, some labels in the
fluid views are always rendered in the default language.

This change sets the preferred browser language to the `BE_USER`
object in the backend login controller and the reset password
controller, when no backend user is logged in, so label translation
is now consistent again.

Resolves: #101166
Related: #100255
Releases: main, 12.4
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: If3ba886c4e8be19990c45cc48c7001a128535f25
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79678
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>