[BUGFIX] Consider more TMENU a-tag related config for runtime cache
Since the swap to LinkResult, the ATagParams and some more attributes
are cached independent of different configuration. This is now fixed by
adding the following configuration options to the md5 hash:

- TMENUITEM: ATagParams
- TMENUITEM: additionalParams
- TMENU: addParams
- TMENU: target

Note: Doing this weakens the usefulness of the runtime cache again
(especially for old-school TypoScript based menus). Another option would
be to add the attributes after the LinkResult generation (how it was
done before), but this would bypass possible Event manipulation.

Resolves: #100319
Related: #96486
Releases: main, 12.4
Change-Id: I8a0346b5b932f071f9498ddccd7ecceca39deed4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79672
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>