[BUGFIX] Do not return exit codes > 255
If a console command throws an exception that isn't caught, the exit
code evaluates to "0" (aka successful) because the exit codes usually
are the exception timestamp and UNIX doesn't support exit codes > 255.

The issue is now solved by checking the exit code in
`CommandApplication` and setting an upper bound of 255.

Resolves: #101220
Releases: main, 12.4
Change-Id: I722125aa678fe24accb7ebf9732e571ff54abb73
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79585
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Susanne Moog <look@susi.dev>