[BUGFIX] Fetch pages in different language when in free mode
The special case for "pages" is that there is always
a record in sys_language_uid=0 and that the pages
in the default language are usually used for linking
and pointing to a page.

For this reason, it is important to actually fetch
the page in language variant "5" for example to also
fetch the page.

Resolves: #101187
Resolves: #101157
Releases: main, 12.4
Change-Id: Ieb7575f0f8c669c1737f55435d1ba991ec8b5019
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79586
Tested-by: Susi Moog <look@susi.dev>
Tested-by: Stefan B�rk <stefan@buerk.tech>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan B�rk <stefan@buerk.tech>
Reviewed-by: Susi Moog <look@susi.dev>