[DOCS] Outline ext:felogin behavior when plugin is hidden at login
This change moves the section about "Access restrictions on the felogin
plugin" to the user manual, since the section is not an example, but
an important notice for users of the plugin. The part about having
two plugins on one page has been removed, since this will lead to
other problems as noted in the "Known problems" chapter.

Additionally, a short notice about the "Hide at login" problem has
been added to the "Known problems" chapter.

Resolves: #100632
Releases: main, 12.4, 11.5
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: I28fe311ce46ef3e1cae9cd60cebb7215fb02701e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79613
Reviewed-by: Jochen <rothjochen@gmail.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>